Revert the changed logic of on_change_with_shipping_date introduced in changeset 804cf9f68f9f3dcd3321f10e0971417acb7abf40.
This change breaks the former behavior and causes a traceback in the workflow on step processing,
because it doesn't providei any shipping date when there is already a move without a delivery_date
(which can be very well the case when the sale line was created from code and not created from the
client with other on_changes being triggered).

Traceback:
TypeError: unsupported operand type(s) for -: 'NoneType' and 'datetime.date'
on processing of a sale line.